Fix some tests after move splash screen window to systeui(11/N)
After splash screen move to systemui, there will be some timing changes
such as add/remove window were not complete synced inside system server.

KeyguardTransitionTests: There will trigger another transition when
remove the starting window, disable showing splash screen for those
tests to make them stable.
Fix testHandleExitIconAnimatingActivity failed due to type cast.

MultiDisplaySystemDecorationTests#testCrossDisplayBasicImeOperations
This test use tapOnDisplaySync to launch mock IME, but the starting
window may above the activity.

Disable starting window for WindowInsetsControllerTests
CloseOnOutsideTests, which may cause test flaky.

Bug: 73289295
Bug: 131311659
Test: atest KeyguardTransitionTests SplashscreenTests
Test: atest MultiDisplaySystemDecorationTests
Test: atest WindowInsetsControllerTests CloseOnOutsideTests

+++ tests/framework/base/windowmanager/src/android/server/wm/SplashscreenTests.java
@@ -241,7 +241,7 @@ public class SplashscreenTests extends ActivityManagerTestBase {
                 () -> journal.extras.getBoolean(RECEIVE_SPLASH_SCREEN_EXIT));
         assertTrue(journal.extras.getBoolean(CONTAINS_CENTER_VIEW));
         final long iconAnimationStart = journal.extras.getLong(ICON_ANIMATION_START);
-        final int iconAnimationDuration = journal.extras.getInt(ICON_ANIMATION_DURATION);
+        final long iconAnimationDuration = journal.extras.getLong(ICON_ANIMATION_DURATION);
         assertTrue(iconAnimationStart != 0);
         assertEquals(iconAnimationDuration, 500);
         assertFalse(journal.extras.getBoolean(CONTAINS_BRANDING_VIEW));
